feat: update ui, refactor
