+++ frontend/app/api/projects/[projectId]/traces/[traceId]/spans/route.ts
@@ -18,7 +18,7 @@ export async function GET(
   );
 
   if (!parseResult.success) {
-    return NextResponse.json({ spans: [] });
+    return NextResponse.json([]);
   }
 
   try {
